refactor: avoid app exports
