+++ server/main.js
@@ -25,6 +25,9 @@ const nodePath = isProduction
 
 (async () => {
   try {
+    const dbLoader = await import('./src/modules/loadDb.js');
+
+    await dbLoader.promise;
     const { startServer, startOSCServer } = await import(nodePath);
     // Start express server
     loaded = await startServer();
