+++ server/src/run.js
@@ -1,6 +1,10 @@
+import { promise } from './modules/loadDb.js';
+
 (async () => {
   let loaded;
   try {
+    await promise;
+
     const { startServer, startOSCServer } = await import('./app.js');
     // Start express server
     loaded = await startServer();
